spunoff server code
Moved server code into sepearate file since mjs does not have access to 'require'. (server file just imports a hardlinked mjs copy of chant to allow for server-side state management using the exact code as the client)
