+++ chant.js
@@ -63,11 +63,10 @@ const chant=function(json={})
 		val=func(util.clone(init));
 		return self.set(path,val);
 	};
-	//client code
 	self.with=function(address=location.href.split('/')[2])//[protocol,_,addr]
 	{
 		//setup socket
-		const socket=sockets[address]=new WebSocket('ws://'+address);
+		const socket=sockets[address]=new WebSocket('ws://'+address,'echo-protocol');
 		//setup state listener
 		self.on({func:function(action)//{type,path,val}
 		{
@@ -92,43 +91,6 @@ const chant=function(json={})
 			console.log('msg received',evt.data);
 		});
 		return self;
-	};
-	//server code
-	self.server=function(httpServer)
-	{
-		const
-		wsServer=require('websocket').server,
-		server=new wsServer({autoAcceptConnections:false,httpServer}),
-		originIsAllowed=origin=>true;//@todo +auth logic
-		server.on('request',function(req)
-		{
-			if (!originIsAllowed(req.origin))
-			{
-				// Make sure we only accept requests from an allowed origin
-				req.reject();
-				console.log((new Date())+' Connection from origin '+req.origin+' rejected.');
-				return;
-			}
-			var connection=req.accept('echo-protocol',req.origin);
-			console.log((new Date())+' Connection accepted.');
-			connection.on('message',function(msg)
-			{
-				if (msg.type==='utf8')
-				{
-					console.log('Received Message: '+msg.utf8Data);
-					connection.sendUTF(msg.utf8Data);
-				}
-				else if (msg.type==='binary')
-				{
-					console.log('Received Binary Message of '+msg.binaryData.length+' bytes');
-					connection.sendBytes(msg.binaryData);
-				}
-			});
-			connection.on('close',function(reasonCode,desc)
-			{
-				console.log((new Date())+' Peer '+connection.remoteAddress+' disconnected.');
-			});
-		});
 	};
 	return self;
 },
